% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.species.info.R
\name{get.species.info}
\alias{get.species.info}
\title{Get info on available species and tissues}
\usage{
get.species.info()
}
\value{
result List containing available species tax ids and names for tissue-specific networks
}
\description{
Get info on available species and tissues
}
\examples{
result = get.species.info()
}
