% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GifyResult.R
\name{GifyResult}
\alias{GifyResult}
\title{GifyResult class containing data from GUILDify results}
\usage{
GifyResult(score.table, function.table, drug.table, enrichment.cutoff, job.id,
  job.id2)
}
\description{
GifyResult class containing data from GUILDify results
}
\keyword{internal}
