% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.job.R
\name{retrieve.job}
\alias{retrieve.job}
\title{Retrieve results using job id}
\usage{
retrieve.job(job.id, n.top = NULL, fetch.files = F, output.dir = NULL)
}
\arguments{
\item{job.id}{Job id}

\item{n.top}{Number of top proteins to retrieve. If NULL top functionally enriched proteins are provided (upto 500 proteins)}

\item{fetch.files}{Flag to fetch result files from server and save them locally in output.dir 
(If NULL files for top functionally enriched proteins are provided, otherwise files for top-ranking 1\% are provided)}

\item{output.dir}{Directory to save the ranking, function, subnetwork and drug info files 
fetched from the server. If NULL, a folder with the same as name as the job.id is created in the current working directory}
}
\value{
result List containing scores of top-ranking proteins, 
        functions enriched among top-ranking proteins,
        drugs targeting top-ranking proteins,
        top-ranking subnetwork cutoff
}
\description{
Retrieve results using job id
}
\examples{
result = retrieve.job(job.id)
getSlots(class(result))
#Scores
head(gScores(result))
#Functions
head(gFunctions(result))
#Drugs
head(gDrugs(result))
}
