% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.overlap.R
\name{retrieve.overlap}
\alias{retrieve.overlap}
\title{Retrieve overlap between two results using two job ids}
\usage{
retrieve.overlap(job.id1, job.id2, top.validated = T, fetch.files = F,
  output.dir = NULL)
}
\arguments{
\item{job.id1}{Job id 1}

\item{job.id2}{Job id 2}

\item{top.validated}{Flag to get overlap between either top-ranking functionally validated proteins (default) or top top-ranking 1\%}

\item{fetch.files}{Flag to fetch result files (for top-ranking 1\%) from server and save them locally in output.dir}

\item{output.dir}{Directory to save the ranking, function, subnetwork and drug info files fetched from the server}
}
\value{
result List containing scores of common top-ranking proteins, 
        common functions enriched among top-ranking proteins,
        drugs targeting common top-ranking proteins
        (Note that the number of top-ranking proteins and common functions are limited to 500)
}
\description{
Retrieve overlap between two results using two job ids
}
\examples{
result = retrieve.overlap(job.id1, job.id2)
getSlots(class(result))
#Scores
head(gScores(result))
#Common functions between top ranking genes of two results
head(gFunctions(result))
#Functions of top ranking common genes of two results
head(gFunctions2(result))
#Drugs
head(gDrugs(result))
}
