% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.job.R
\name{submit.job}
\alias{submit.job}
\title{Submit job using the guildifyR query result table}
\usage{
submit.job(result.table, species, tissue, scoring.options = list(netcombo =
  T))
}
\arguments{
\item{scoring.options}{Prioritization method options (default NetCombo)}

\item{query.table}{Data frame contaning information from guildifyR query (user entitiy info, network info, source etc...)}
}
\value{
job id
}
\description{
Submit job using the guildifyR query result table
}
\examples{
species="10090"
tissue="All"
result.table = query("alzheimer", species)
#Diamond
scoring.options = list(diamond=T)
#NetScore
scoring.options = list(netscore=T, repetitionSelector=3, iterationSelector=2)
#NetZcore
scoring.options = list(netzcore=T, repetitionZelector=3)
#NetShort
scoring.options = list(netshort=T)
#NetCombo
scoring.options = list(netcombo=T)
job.id = submit.job(result.table, species, tissue, scoring.options)
}
