% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.parameters.R
\name{check.parameters}
\alias{check.parameters}
\title{Check whether provided species, tissue and scoring parameter values are valid}
\usage{
check.parameters(species, tissue, scoring.options = NULL)
}
\description{
Check whether provided species, tissue and scoring parameter values are valid
}
