% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Get protein/gene info associated with the query keywords}
\usage{
query(keywords, species = "9606", tissue = "All", whitespaces.as.and = T)
}
\arguments{
\item{keywords}{Text containing description of a phenotype or list of genes (seperated by ;)}

\item{species}{Species tax identifier (9606: human, 10090: mouse, etc., see get.species.info method)}

\item{whitespaces.as.and}{Treat whitespaces as ANDs in querying (quotes the keywords)}

\item{tissues}{Tissue identifier (All, brain, liver, etc., see get.species.info method)}
}
\value{
result.table Data frame containing list of matching proteins/genes and their description
}
\description{
Get protein/gene info associated with the query keywords
}
\examples{
result.table = query("alzheimer", species="10090", tissue="All")
}
