% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Get protein/gene info associated with the query keywords}
\usage{
query(keywords, species = "9606", tissue = "all",
  network.source = "BIANA", quote.keywords = T)
}
\arguments{
\item{keywords}{Text containing description of a phenotype or list of genes (seperated by ";")}

\item{species}{Species tax identifier (9606: human, 10090: mouse, etc., see get.species.info method)}

\item{tissue}{Tissue identifier (all, liver, lung, etc., see get.species.info method)}

\item{network.source}{Interaction network source (BIANA, STRING, HIPPIE, etc.)}

\item{quote.keywords}{Quotes the keywords to treat whitespaces as ANDs (e.g., "Muscular Dystrophy")}
}
\value{
result.table Data frame containing list of matching proteins/genes and their description
}
\description{
Get protein/gene info associated with the query keywords
}
\examples{
result.table = query("alzheimer", species="10090", tissue="all", network.source="BIANA")
}
