% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.job.R
\name{retrieve.job}
\alias{retrieve.job}
\title{Retrieve results using job id}
\usage{
retrieve.job(job.id, n.top = NULL, fetch.files = F, output.dir = "./")
}
\arguments{
\item{job.id}{Job id}

\item{n.top}{Number of top proteins to retrieve. If NULL top enriched proteins are provided (upto 500 proteins)}

\item{fetch.files}{Flag to fetch result files from server and save them locally in output.dir}

\item{output.dir}{Directory to save the ranking, function, subnetwork and drug info files fetched from the server}
}
\value{
list containing scores of top-ranking proteins, 
        functions enriched among top-ranking proteins,
        drugs targeting top-ranking proteins,
        top-ranking subnetwork cutoff
}
\description{
Retrieve results using job id
}
\examples{
result = retrieve.job(job.id)
names(result)
head(result$score.table)
}
