% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.overlap.R
\name{retrieve.overlap}
\alias{retrieve.overlap}
\title{Retrieve overlap between two results using two job ids}
\usage{
retrieve.overlap(job.id1, job.id2, fetch.files = F, output.dir = "./")
}
\arguments{
\item{job.id1}{Job id 1}

\item{job.id2}{Job id 2}

\item{fetch.files}{Flag to fetch result files from server and save them locally in output.dir}

\item{output.dir}{Directory to save the ranking, function, subnetwork and drug info files fetched from the server}
}
\value{
list containing scores of common top-ranking proteins, 
        common functions enriched among top-ranking proteins,
        drugs targeting common top-ranking proteins
        (Note that the number of top-ranking proteins and common functions are limited to 500)
}
\description{
Retrieve overlap between two results using two job ids
}
\examples{
result = retrieve.job(job.id)
names(result)
head(result$score.table)
}
